Move observable infrastructure out of wizard module
