+++ src/components/wizard/wizard.ts
@@ -1,24 +1,7 @@
 import * as vscode from 'vscode';
+import { Sequence, isObservable, isThenable } from '../../utils/observable';
 
 export const NEXT_FN = "onNext();";
-
-export interface Observable<T> {
-    subscribe(observer: Observer<T>): void;
-}
-
-export interface Observer<T> {
-    onNext(value: T): Promise<boolean>;
-}
-
-export type Sequence<T> = T | Thenable<T> | Observable<T>;
-
-function isObservable<T>(s: Sequence<T>): s is Observable<T> {
-    return !!((s as Observable<T>).subscribe);
-}
-
-function isThenable<T>(s: Sequence<T>): s is Thenable<T> {
-    return !!((s as Thenable<T>).then);
-}
 
 export class Wizard {
     constructor(private readonly w: vscode.WebviewPanel) { }
